cleaup

+++ src/main.c
@@ -36,12 +36,12 @@ BYTE isPositionWhite() {
 
 // https://archive.org/details/The_Graphics_Book_for_the_Commodore_64/page/n129/
 void setPositionWhite() {
-    ad = 0x2000+(320 * (y/8)) + (y & 7)+8 * (short)(x/8);
+    ad = 0x2000+(320 * (y/8)) + (y & 7)+8 * (x/8);
     *(short*)(ad) = *(short*)(ad) | 1 << ((7-(x & 7)));
 }
 
 void setPositionBlack() {
-    ad = 0x2000+320 * (y/8) + (y & 7)+8 * (short)(x/8);
+    ad = 0x2000+320 * (y/8) + (y & 7)+8 * (x/8);
     *(short*)(ad) = (*(short*)(ad)) & ~(1 << ((7-(x & 7))));
 }
